Bootstrap for simulation
The following patch adds the necessary bootstrapping for the simulation.

As the saying goes: "everything is a nail when you have a hammer".
Pytest is usually used for testing, but it can be honestly used as a runner
for almost everything. The scoping and lifetime of the fixtures,
combined with the testcase isolation and parametrization, makes
writing a simulation a breeze.

One pending thing is to better visualize the values and hopefully
we see some nice trends.